requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
discord.py==2.4.0
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
multidict==6.1.0
openai==1.47.0
pydantic==2.9.2
pydantic-core==2.23.4
pytelegrambotapi==4.23.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
youtube-transcript-api==0.6.2
